requirements_dev.txt
pip==18.1
bumpversion==0.5.3
wheel==0.32.1
watchdog==0.9.0
flake8==3.5.0
tox==3.5.2
coverage==4.5.1
Sphinx==1.8.1
twine==1.12.1
atomicwrites==1.2.1
gevent==1.2.2
attrs==18.2.0
bottle==0.12.13
bottle-websocket==0.2.9
certifi==2018.10.15
chardet==3.0.4
future==0.17.1
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.7
more-itertools==4.3.0
mypy==0.641
mypy-extensions==0.4.1
pluggy==0.8.0
pytest==4.0.0
pytest-base-url==1.4.1
pytest-html==1.19.0
pytest-metadata==1.7.0
pytest-selenium==1.14.0
pytest-variables==1.7.1
selenium==3.141.0
py==1.7.0
requests==2.20.1
six==1.11.0
Transcrypt==3.7.11
typed-ast==1.1.0
urllib3==1.24.1
whichcraft==0.5.2
